update: moved all styles in add user model to a new style file
